Add combined substance quantities

Adds use days with combined quantities for marijuana and/or nicotine by combining all individually reported substance quantities for the respective substance class on a given day and applying standard factors to convert them into the same unit.

Usage

add_combined_substances(data, substance, co_use = NULL)

Arguments

data

tibble. A data frame containing the TLFB raw data.

substance

character (vector). The substance(s) to compute the score for. Must be one or several of the following values:

"'Fake' Marijuana or Synthetics"

"Alcohol"

"Anabolic Steroids"

"Any Other Drug They Used to Get High"

"Blunts or Combined Tobacco and Marijuana in Joints"

"CBD (Non-Medical Use)"

"Cathinones such as Bath Salts, Drone, or Meph"

"Cigars, Little Cigars, or Cigarillos"

"Cocaine or Crack Cocaine"

"Concentrated Marijuana Tinctures"

"Ecstasy, Molly, or MDMA"

"Electronic Nicotine or Vaping Products"

"GHB, Liquid G, or Georgia Homeboy"

"Hallucinogen Drugs including LSD, PCP, Peyote, Mescaline, DMT, AMT, or Foxy"

"Heroin, Opium, Junk, Smack, or Dope"

"Hookah with Tobacco"

"Inhalants"

"Ketamine or Special K"

"Marijuana Edibles"

"Marijuana Infused Alcohol Drinks"

"Methamphetamine, Meth, or Crystal Meth"

"Nicotine Replacements"

"OTC Cough or Cold Medicine, DXM, 'Lean', or 'Purple Drank'"

"Prescription Anxiolytics, Tranquilizers, or Sedatives"

"Prescription Pain Relievers or Opioids"

"Prescription Stimulants"

"Psilocybin, Magic Mushrooms, or Shrooms"

"Salvia"

"Smokeless Tobacco, Chew, or Snus"

"Smoking Marijuana Flower"

"Smoking Marijuana Oils or Concentrates"

"Tobacco Cigarette"

"Tobacco in a Pipe"

"Vaped Marijuana Flower"

"Vaped Marijuana Oils or Concentrates"

"Marijuana (all forms)"

"Nicotine (all forms)"

(Default: NULL, i.e., all substances are considered.)

co_use

character (vector). Co-use substance(s). Must be one or several of the possible values for substance listed above. Only days where the specified substance(s) was/were used together with (one of) the co-use substance(s) are considered. (Default: NULL, i.e., co-use is not considered). co_use cannot be specified without substance and can only contain substance(s) that are not specified in substance.

Value

A data frame with the original data and, if substance contains either one of "Marijuana (all forms)" or "Nicotine (all forms)", newly created combined substance(s) use days and quantities.

Examples

if (FALSE) { # \dontrun{
data <- tibble::tibble(
  substance = c(
    "Smoking Marijuana Flower",
    "Vaped Marijuana Oils or Concentrates",
    "Tobacco Cigarette"
  ),
  quantity = c(10, 5, 20)
)
add_combined_substances(data, "Marijuana (all forms)")
add_combined_substances(data, "Smoking Marijuana Flower")
} # }