Compute "Family Environment Scale [Parent] (Expression): Number missing"

Computes the summary score fc_p_fes__expr_nm (Family Environment Scale [Parent] (Expression): Number missing)

Summarized variables:

fc_p_fes__expr_001

fc_p_fes__expr_002

fc_p_fes__expr_003

fc_p_fes__expr_004

fc_p_fes__expr_005

fc_p_fes__expr_006

fc_p_fes__expr_007

fc_p_fes__expr_008

fc_p_fes__expr_009

Excluded values: none

Usage

compute_fc_p_fes__expr_nm(data, name = "fc_p_fes__expr_nm", combine = TRUE)

Arguments

data

tbl. Data frame containing the columns to be summarized.

name

character. Name of the summary score. Default is the name in the description.

combine

logical. If TRUE, the summary score is appended to the input data frame. If FALSE, the summary score is returned as a separate data frame. Default is TRUE.

Value

tbl. The input data frame with the summary score appended as a new column.

See also