Compute all the fc_y_wpss summary scores

This is a high-level function that computes all summary scores in this table. Make sure the data contains all the necessary columns.

Usage

compute_fc_y_wpss_all(data)

Arguments

data

tbl, Dataframe containing the columns to be summarized.

Value

tbl. The input data frame with the summary scores appended as new columns.

Examples

if (FALSE) { # \dontrun{
compute_fc_y_wpss_all(data)
} # }