Compute "Early Adolescent Temperament Questionnaire [Parent] (Affiliation): Number missing"

Computes the summary score mh_p_eatq__affl_nm Early Adolescent Temperament Questionnaire [Parent] (Affiliation): Number missing

Summarized variables:

mh_p_eatq__affl_001

mh_p_eatq__affl_002

mh_p_eatq__affl_003

mh_p_eatq__affl_004

mh_p_eatq__affl_005

mh_p_eatq__affl_006

Usage

compute_mh_p_eatq__affl_nm(data, name = "mh_p_eatq__affl_nm", combine = TRUE)

Arguments

data

tbl, Dataframe containing the columns to be summarized.

name

character, Name of the new column to be created. Default is the name in description, but users can change it.

combine

logical, If TRUE, the summary score will be appended to the input data frame. If FALSE, the summary score will be returned as a separate data frame.

Value

tbl. The input data frame with the summary score appended as a new column.

See also

Examples

if (FALSE) { # \dontrun{
data <- compute_mh_p_eatq__affl_nm(data)
select(
  data,
  any_of(c("mh_p_eatq__affl_nm", vars_mh_p_eatq__affl))
)
} # }